error-reporter: Move server error reporter into its own top level module
Change-type: patch
