+++ lib/error-reporter/index.js
