+++ lib/server/action-server.js
@@ -21,7 +21,7 @@ const _ = require('lodash')
 const Worker = require('../worker')
 const logger = require('../logger').getLogger(__filename)
 const actionLibrary = require('../action-library')
-const errorReporter = require('./error-reporter')
+const errorReporter = require('../error-reporter')
 
 let run = true
