+++ lib/server/index.js
@@ -16,9 +16,9 @@
 
 const randomstring = require('randomstring')
 const errio = require('errio')
+const errorReporter = require('../error-reporter')
 const logger = require('../logger').getLogger(__filename)
 const bootstrap = require('./bootstrap')
-const errorReporter = require('./error-reporter')
 
 const context = {
 	id: `SERVER-${randomstring.generate(20)}`
